// src/app/models/collaborator.model.ts
export class CollaboratorModel {

    id: number = 0;    
    collaboratorName: string = "";  
    collaboratorLastName: string = "";
    collaboratorPosition: string = "";
    specialityArea: string = "";      
    documentCategory: string = "";      
    identificationDocument: number = 0;     
    state: number = 0;     
}